refactor httpRequest
